+++ src/main/java/fll/web/admin/TeamSchedules.java
@@ -13,7 +13,9 @@ import java.util.Set;
 
 import javax.sql.DataSource;
 
-import fll.db.Queries;
+import fll.Team;
+import fll.Tournament;
+import fll.scheduler.SchedParams;
 import fll.scheduler.ScheduleWriter;
 import fll.scheduler.TournamentSchedule;
 import fll.web.ApplicationAttributes;
@@ -32,6 +34,8 @@ import jakarta.servlet.http.HttpSession;
 /**
  * @see ScheduleWriter#outputTeamSchedules(TournamentSchedule,
  *      fll.scheduler.SchedParams, java.io.OutputStream)
+ * @see ScheduleWriter#outputTeamSchedule(TournamentSchedule, SchedParams,
+ *      java.io.OutputStream, int)
  */
 @WebServlet("/admin/TeamSchedules")
 public class TeamSchedules extends BaseFLLServlet {
@@ -52,21 +56,29 @@ public class TeamSchedules extends BaseFLLServlet {
 
     final DataSource datasource = ApplicationAttributes.getDataSource(application);
     try (Connection connection = datasource.getConnection()) {
+      final Tournament tournament = Tournament.getCurrentTournament(connection);
 
-      final int currentTournamentID = Queries.getCurrentTournament(connection);
-
-      if (!TournamentSchedule.scheduleExistsInDatabase(connection, currentTournamentID)) {
+      if (!TournamentSchedule.scheduleExistsInDatabase(connection, tournament.getTournamentID())) {
         SessionAttributes.appendToMessage(session, "<p class='error'>There is no schedule for this tournament.</p>");
         WebUtils.sendRedirect(application, response, "/admin/index.jsp");
         return;
       }
 
-      final TournamentSchedule schedule = new TournamentSchedule(connection, currentTournamentID);
+      final TournamentSchedule schedule = new TournamentSchedule(connection, tournament.getTournamentID());
+      final SchedParams schedParams = new SchedParams();
+      schedParams.load(connection, tournament, schedule);
 
       response.reset();
       response.setContentType("application/pdf");
-      response.setHeader("Content-Disposition", "filename=teamSchedules.pdf");
-      ScheduleWriter.outputTeamSchedules(schedule, schedParams, response.getOutputStream());
+
+      final int teamNumber = WebUtils.getIntRequestParameter(request, "TeamNumber", Team.NULL_TEAM_NUMBER);
+      if (Team.NULL_TEAM_NUMBER == teamNumber) {
+        response.setHeader("Content-Disposition", "filename=teamSchedules.pdf");
+        ScheduleWriter.outputTeamSchedules(schedule, schedParams, response.getOutputStream());
+      } else {
+        response.setHeader("Content-Disposition", String.format("filename=teamSchedule-%d.pdf", teamNumber));
+        ScheduleWriter.outputTeamSchedule(schedule, schedParams, response.getOutputStream(), teamNumber);
+      }
 
     } catch (final SQLException sqle) {
       LOGGER.error(sqle.getMessage(), sqle);